Stage 1/5: Putting two and two together
Description

In the first stage, we're going to create a simple version of our calculator. It is going to become more advanced and complex in the upcoming stages.

You should add a single JButton and a single JTextField to the main program window. A click on the button adds both numbers in the JTextField and displays the result in the format 2+2 => 2+2=4, where 2+2 is the equation, and 2+2=4 is the result.

In this project, we are not going to use any layout managers, so it's up to you to customize the calculator look.

Objectives
In this stage, you should:

Set the title of the window as Calculator;
Add a JTextField to the main window;
Add a JButton to the main window;
Set the names of the following components:
JTextField as EquationTextField;
JButton as Solve.