Stage 2/5: Making buttons
Description

In the previous stage, we have created a very simple calculator that can add two numbers, but let's face it — it's boring. So in this stage, we are going to expand our calculator to support subtraction, division, and multiplication. We will also take our first steps in designing our calculator.

As you may have guessed, we will need buttons, lots of buttons.

Add buttons for numbers (0 to 9), for each operator (addition, subtraction, division, multiplication), and of course, for the equals = sign. A click on a button should add the corresponding character to the JTextField, except the equals button that should perform the corresponding operation on both numbers (add a to b, subtract b from a and so on).

Objectives

In this stage, you should:

1. Add JButtons to the window for numbers;
2. Add JButtons to the window for operators;
3. Add a JButton to the window for the equals sign;
4. Set the names of the following components:
  JTextField as EquationTextField
  10 JButtons for numbers from Zero to Nine;
  4 JButtons as Add, Subtract, Divide, Multiply;
  1 JButton as Equals.