Stage 3/5: Operator precedence
Description

So far, our calculator can perform four basic operations with two numbers. We can do better. In this stage, we will improve our calculator so that it will be able to perform different operations on different sets of numbers. We will also work on the design of our calculator. As of now, it displays the equation and the result in a single JTextField. We will replace it with two JLabels.

Let's talk briefly about the operator precedence. In math, some operators have a higher weight than others. As you probably remember, division and multiplication have higher precedence than addition and subtraction. If an equation has two or more operators with equal precedence, they are executed from left to right. Take for example the equation 5+4-3x2/1. We get => 5+4-6/1 => 5+4-6 => 9-6 == 3. We know that multiplication and division have equal precedence, but multiplication was executed first, as it was the first in the equation. Then, we divide, add and, finally, subtract. We need to account for this in our app.

To evaluate each expression, you may need to convert it from infix notation to postfix notation. Infix notation is how we usually take down expressions. Each operator follows an operand. The expressions above can serve as examples of infix notation. Another type of notation is postfix. In this case, operators are followed by operands. For example, the postfix notation of the expression 5+4-3 is 5 4 3 + -.

Replace the JTextField with two JLabels, one for storing the equation and one for displaying the result. Also, add three new JButtons, a decimal button that places a decimal point in a JLabel equation, a CLEAR button that clears all characters, and a delete button.

For testing reasons, use the Unicode format for operators:

Addition — \u002B
Subtraction — \u2212
Multiplication — \u00D7
Division — \u00F7

Objectives
In this stage, you should:

Add two JLabels — one for storing an equation, the other for displaying the result;
Add a JButton for inserting a decimal point .;
Add a JButton for clearing an equation;
Add a JButton for deleting the last character in an equation;
Set the names for the following components:
JLabel as EquationLabel;
JLabel as ResultLabel;
10 JButtons for numbers from Zero to Nine;
4 JButtons as Add, Subtract, Divide, Multiply;
JButton as Equals;
JButton as Clear;
JButton as Delete;
JButton as Dot;