Stage 5/5: Advanced features
Description

In the final stage, we will add more functions to our calculator. We have added the ability to evaluate expressions of variable length and operator precedence before. Now we will override the precedence using parentheses. We will also add the square root function, exponentiation (X raised to the power of 2 and X raised to the power of Y, where X and Y are real numbers), and, finally, negation.

Let's cover them step by step.

Parentheses
Add a single JButton that will insert both left ( and right ) parentheses. There are a few cases that determine which parenthesis is to be inserted. Insert a left parenthesis if the number of left parentheses is equal to the number of right parentheses in the expression. Insert a left parenthesis if the last character in the expression is another left bracket, or if the last character in the expression is an operator. Otherwise, insert a right parenthesis. The expression in the parentheses should be calculated first.

Square Root
Add a new JButton for the square root function. When the square root button is clicked, insert a radical symbol \(√\) followed by a left parenthesis. The result of the expression in the parentheses will be used as input to the square root function.

The Unicode for the square root symbol is \u221A.

Exponentiation
Add a JButton for the Raise to 2 function and another JButton for the Raise to Y function. A click on the Raise to 2 button inserts ^ as an operator followed by (2), so the number on the left side of the operator is raised to the power of 2. Similarly, a click on the Raise to Y button inserts a power operator ^ followed by a left parenthesis to allow for any real number or expression, which serves as input.

Negation
Add a JButton to negate a number or an expression. A click on this button inserts a left parenthesis followed by the subtract symbol (- that will negate any number that comes after it. Note that clicking on this button should consecutively undo the negation. For instance, if the expression is empty, the first click on this button inserts (-. The second click removes (- from the expression. We know that having (- before a number negates it, but what about negating a full expression? To do it, we need to have (- and an expression in another pair of parentheses. For instance, the negative version of 2+4-7 is (-(2+4-7)).

The guidelines for validating and formatting input established in the previous stage still hold true, except for two things. Square root can start an expression, and you cannot get a negative number by squaring since a square is either positive or 0.

Objectives
In this stage, you should:

Add a JButton for handling left and right parentheses;
Add a JButton for the square root function;
Add a JButton for the Raise to 2 function;
Add a JButton for the Raise to Y function;
Add a JButton for negating a number or an expression.
Maintain the names of the interface components (JLabels and JButtons) from the previous stage. For new elements, set the following names:
JButton as Parentheses;
JButton as SquareRoot;
JButton as PowerTwo;
JButton as PowerY;
JButton as PlusMinus.